[fix]controller 파일 수정
